change REPEATS from numeric/double to integer/int
git-svn-id: https://svn.code.sf.net/p/moneymanagerex/code/trunk@2462 7f5d68d0-c712-0410-839f-9b116a576cc5

+++ mmex/sql/tables_v1.sql
@@ -44,7 +44,7 @@ BDID integer primary key
 , TRANSDATE TEXT
 , FOLLOWUPID integer
 , TOTRANSAMOUNT numeric
-, REPEATS numeric
+, REPEATS integer
 , NEXTOCCURRENCEDATE TEXT
 , NUMOCCURRENCES numeric
 );
